Move test helper methods to utility dir as functions.  Deprecate containing class.
