+++ src/Mock/module.ts
@@ -1,0 +1,2 @@
+export * from './Mock';
+export * from './Times';
