+++ src/public_api.ts
@@ -1,8 +1,2 @@
-/*
- * Public API Surface of tsmockit
- */
-
-export * from './Mock/Mock';
-export * from './Mock/Times';
-
-export * from './Utility/TestHelpers';
+export * from './Mock/module';
+export * from './Utility/module';
